Добавляет загрузку избранного при запуске приложения
