+++ src/pages/offer/offer.tsx
@@ -7,7 +7,6 @@ import Host from '../../components/host/host';
 import Map from '../../components/map/map';
 import { useParams } from 'react-router-dom';
 import NotFound from '../not-found-page/not-found';
-import type { OfferPage } from '../../types/types';
 import { Helmet } from 'react-helmet-async';
 import FavoriteButton from '../../components/favorite-button/favorite-button';
 import { getMarkupRating } from '../../utils/common';
@@ -16,7 +15,7 @@ import { AuthorizationStatus, MAX_NEARBY_OFFERS_COUNT, MAX_IMAGES_COUNT } from '
 import ReviewForm from '../../components/review-form/review-form';
 import PlaceCardList from '../../components/place-card-list/place-card-list';
 import { useAppDispatch, useAppSelector } from '../../hooks';
-import { fetchOfferDetailsById, fetchOfferComments, fetchNearbyOffers, fetchFavoriteOffers } from '../../store/thunk-actions';
+import { fetchOfferDetailsById, fetchOfferComments, fetchNearbyOffers } from '../../store/thunk-actions';
 import { selectCity } from '../../store/active-city/active-sity-selector';
 import { selectAuthorizationStatus } from '../../store/auth/auth-selector';
 import { selectNearbyOffers, selectOfferDetails, selectOfferComments } from '../../store/offers/offer-selector';
@@ -41,13 +40,6 @@ function OfferPage(): JSX.Element {
   const currentOffer = useAppSelector(selectOfferDetails);
   const reviews = useAppSelector(selectOfferComments);
   const nearOffers = useAppSelector(selectNearbyOffers);
-
-
-  useEffect(() => {
-    if (authorizationStatus === AuthorizationStatus.Auth) {
-      dispatch(fetchFavoriteOffers());
-    }
-  }, [dispatch, authorizationStatus]);
 
   if (!currentOffer) {
     return <NotFound />;
